ログイン画面の追加修正、管理者プロフィール画面、管理者管理画面の追加実装

+++ ProductInventorySystem/src/main/java/com/example/demo/Controller/ManagerController.java
@@ -15,7 +15,7 @@ public class ManagerController {
     private ManagerRepository managerRepository;
 
     @GetMapping("/api/managers/{manager_id}")
-    public Manager getManagerById(@PathVariable("manager_id") Long id) { // 引数をLongに変更
+    public Manager getManagerById(@PathVariable("manager_id") Integer id) { // 引数をIntegerに変更
         return managerRepository.findById(id).orElse(null);
     }
 }
